refactoring e2e-tests
