+++ tests/e2e/matches.spec.ts
@@ -1,0 +1,14 @@
+/**
+ * @file A basic set of E2E-tests running in a real browser environment.
+ * ---
+ * Since modern browser environments already have native support
+ * for the `:has` pseudo-class, for the correct testing of the polyfill,
+ * the `:has` selector is intentionally redefined to `:_has`.
+ *
+ * The result of the polyfill's work is then compared with the native result
+ * in the browser.
+ */
+
+import {expect} from '@playwright/test';
+import {SelectorHandler} from '../../src/';
+import {test} from './global';
